+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # GOJI Changelog
 
-## Master
+## 0.4.0
 
 ### Breaking Changes